feat: 날짜 객체를 시간 문자열로 바꿔주는 유틸함수 작성
