+++ frontend/src/utils/date/time.ts
@@ -1,0 +1,8 @@
+export const formatDateToTimeString = (date: Date): string => {
+  const hours = date.getHours().toString()
+    .padStart(2, '0');
+  const minutes = date.getMinutes().toString()
+    .padStart(2, '0');
+
+  return `${hours}:${minutes}`;
+};
